Update sample code to use Boolean operator
Combining bit flags should be done using the Boolean operator 'Or', rather than addition (+).
In this case, the result is the same, but in some cases addition will produce incorrect results when trying to set a bit (e.g. if adding to a variable that already has the bit set), so our sample code should set a good example.